+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 -   Fix: [Symbol is resolving to efun instead of sefun #95](https://github.com/jlchmura/lpc-language-server/issues/95)
 -   Fix: [Symbols with super prefix do not resolve #97](https://github.com/jlchmura/lpc-language-server/issues/97)
 -   Added ability to use `@callback` and `@template` tags in function signatures.
+-   Properly resolve call expression to a function symbol when a block scoped variable by the same name also exists.
 -   Enhance efun definitions
 
 ## 1.1.7